fixed problem with discovery dialog
- closes #2089
