+++ src/src/dialogs/GenerateInputsModal.jsx
